changes for new sdk 1.5
