+++ sample-solutions/VisionAiDevKit/.env
@@ -1,0 +1,5 @@
+# Replace the value of these variables with your own container registry
+
+CONTAINER_REGISTRY_ADDRESS=""
+CONTAINER_REGISTRY_USERNAME=""
+CONTAINER_REGISTRY_PASSWORD=""
