+++ sample-solutions/VisionAiDevKit/modules/VisionSampleModule/Dockerfile.arm32v7
@@ -36,9 +36,9 @@ COPY . .
 
 RUN chmod -R 777 /app/model
 
-RUN pip3 install -e /app/python_iotcc_sdk
+#RUN pip3 install -e /app/python_iotcc_sdk
 
-ENTRYPOINT [ "python3", "-u", "./python_iotcc_sdk/tests/test-main.py"]
+ENTRYPOINT [ "python3", "-u", "./python_iotcc_sdk/sdk/main.py"]
 # Sets the flag "TRUE" to send dlc file to device else False
 CMD ["-p True"]
