+++ sample-solutions/VisionAiDevKit/modules/VisionSampleModule/model/coco_labels.txt
@@ -1,4 +1,4 @@
-person
+unlabeled
 person
 bicycle
 car
