--- sample-solutions/VisionAiDevKit/modules/VisionSampleModule/python_iotcc_sdk/iotccsdk.egg-info/PKG-INFO
@@ -1,21 +1,0 @@
-Metadata-Version: 1.1
-Name: iotccsdk
-Version: 0.1.0+dev
-Summary: IOTCC SDK
-Home-page: UNKNOWN
-Author: UNKNOWN
-Author-email: UNKNOWN
-License: BSD 2.0
-Description: Python IOTCC Software Development Kit
-Keywords: iot,ai,camera
-Platform: UNKNOWN
-Classifier: Development Status :: 4 - Beta
-Classifier: Framework :: IPython
-Classifier: Intended Audience :: Developers
-Classifier: Intended Audience :: Science/Research
-Classifier: Topic :: Multimedia :: Graphics
-Classifier: Programming Language :: Python :: 3
-Classifier: Programming Language :: Python :: 3.3
-Classifier: Programming Language :: Python :: 3.4
-Classifier: Programming Language :: Python :: 3.5
-Classifier: Programming Language :: Python :: 3.6
