--- sample-solutions/VisionAiDevKit/modules/VisionSampleModule/python_iotcc_sdk/iotccsdk.egg-info/SOURCES.txt
@@ -1,13 +1,0 @@
-README.md
-setup.cfg
-iotcc/__init__.py
-iotcc/sdk/__init__.py
-iotcc/sdk/camera.py
-iotcc/sdk/frame_iterators.py
-iotcc/sdk/ipcprovider.py
-iotcc/sdk/model.py
-iotccsdk.egg-info/PKG-INFO
-iotccsdk.egg-info/SOURCES.txt
-iotccsdk.egg-info/dependency_links.txt
-iotccsdk.egg-info/not-zip-safe
-iotccsdk.egg-info/top_level.txt
