--- sample-solutions/VisionAiDevKit/modules/VisionSampleModule/python_iotcc_sdk/iotccsdk.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-iotcc
