+++ sample-solutions/VisionAiDevKit/modules/VisionSampleModule/python_iotcc_sdk/setup.cfg
