+++ sample-solutions/VisionAiDevKit/modules/VisionSampleModule/python_iotcc_sdk/setup.py
@@ -1,6 +1,29 @@
-# Copyright (c) 2018, The Linux Foundation. All rights reserved.
-# Licensed under the BSD License 2.0 license. See LICENSE file in the project root for
-# full license information.
+# Copyright (c) 2018-2019, The Linux Foundation. All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are
+# met:
+#    * Redistributions of source code must retain the above copyright
+#      notice, this list of conditions and the following disclaimer.
+#    * Redistributions in binary form must reproduce the above
+#      copyright notice, this list of conditions and the following
+#      disclaimer in the documentation and/or other materials provided
+#      with the distribution.
+#    * Neither the name of The Linux Foundation nor the names of its
+#      contributors may be used to endorse or promote products derived
+#      from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED "AS IS" AND ANY EXPRESS OR IMPLIED
+# WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT
+# ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS
+# BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
+# BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+# WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
+# OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
+# IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from setuptools import setup, find_packages
 import os
@@ -8,8 +31,7 @@ import shutil
 
 _major = '0.0'
 _minor = '0+dev'
-
-shutil.copy('.license', 'LICENSE.txt')
+_packages = find_packages(exclude=["*.tests"])
 
 if os.path.exists('major.version'):
     with open('major.version', 'rt') as bf:
@@ -25,7 +47,7 @@ NAME = "iotccsdk"
 LONG_DESCRIPTION = 'Python IOTCC Software Development Kit'
 DESCRIPTION = 'IOTCC SDK'
 
-DEPENDENCIES = [ ]
+DEPENDENCIES = ["pip >= 9.0.0", "requests", "setuptools-git"]
 
 setup_args = {
     'name': NAME,
@@ -33,10 +55,12 @@ setup_args = {
     'description': DESCRIPTION,
     'long_description': LONG_DESCRIPTION,
     'include_package_data': True,
+    'python_requires': '~=3.5',
     'install_requires': [
         DEPENDENCIES
     ],
-    'packages': find_packages(exclude=["*.tests"]),
+    'packages': _packages,
+    'package_data': {'sdk': 'logger.conf'},
     'zip_safe': False,
     'cmdclass': {
     },
@@ -45,18 +69,14 @@ setup_args = {
         'ai',
         'camera',
     ],
-    'license': 'BSD 2.0',
+    'license': 'BSD-3-Clause',
     'classifiers': [
         'Development Status :: 4 - Beta',
         'Framework :: IPython',
         'Intended Audience :: Developers',
         'Intended Audience :: Science/Research',
         'Topic :: Multimedia :: Graphics',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.3',
-        'Programming Language :: Python :: 3.4',
         'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
     ],
 }
